+++ RedPandaIDE/settingsdialog/editorgeneralwidget.cpp
@@ -38,7 +38,6 @@ void EditorGeneralWidget::doLoad()
 {
     pSettings->editor().load();
     //indents
-    ui->chkAddIndent->setChecked(pSettings->editor().addIndent());
     ui->chkAutoIndent->setChecked(pSettings->editor().autoIndent());
     ui->chkTabToSpaces->setChecked(pSettings->editor().tabToSpaces());
     ui->spTabWidth->setValue(pSettings->editor().tabWidth());
@@ -70,7 +69,6 @@ void EditorGeneralWidget::doLoad()
 void EditorGeneralWidget::doSave()
 {
     //indents
-    pSettings->editor().setAddIndent(ui->chkAddIndent->isChecked());
     pSettings->editor().setAutoIndent(ui->chkAutoIndent->isChecked());
     pSettings->editor().setTabToSpaces(ui->chkTabToSpaces->isChecked());
     pSettings->editor().setTabWidth(ui->spTabWidth->value());